Restructure spec files into hierarchical when_*/and_* folders
Group related and_* test classes under their parent when_* folder so the folder
tree mirrors the BDD naming hierarchy instead of having a flat list of long filenames.
