+++ Source/AuthProxy.Specs/Invites/for_InviteMiddleware/when_authenticated_user_has_pending_invite/and_exchange_call_throws.cs
@@ -1,14 +1,10 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Net;
+namespace Cratis.AuthProxy.Invites.for_InviteMiddleware.when_authenticated_user_has_pending_invite;
 
-namespace Cratis.AuthProxy.Invites.for_InviteMiddleware;
-
-public class when_authenticated_user_has_pending_invite_and_lobby_is_configured : Specification
+public class and_exchange_call_throws : Specification
 {
-    const string LobbyUrl = "http://lobby-service/";
-
     InviteMiddleware _middleware;
     DefaultHttpContext _context;
     bool _nextCalled;
@@ -24,7 +20,7 @@ public class when_authenticated_user_has_pending_invite_and_lobby_is_configured 
                 ExchangeUrl = "http://studio/internal/invites/exchange",
                 Lobby = new C.Service
                 {
-                    Frontend = new C.ServiceEndpoint { BaseUrl = LobbyUrl }
+                    Frontend = new C.ServiceEndpoint { BaseUrl = "http://lobby-service/" }
                 }
             }
         };
@@ -32,8 +28,7 @@ public class when_authenticated_user_has_pending_invite_and_lobby_is_configured 
         optionsMonitor.CurrentValue.Returns(config);
 
         var httpClientFactory = Substitute.For<IHttpClientFactory>();
-        httpClientFactory.CreateClient(Arg.Any<string>()).Returns(
-            new HttpClient(new FakeHttpMessageHandler(HttpStatusCode.OK)));
+        httpClientFactory.CreateClient(Arg.Any<string>()).Returns(new HttpClient(new ThrowingHttpMessageHandler()));
 
         _middleware = new InviteMiddleware(
             _ =>
@@ -51,23 +46,27 @@ public class when_authenticated_user_has_pending_invite_and_lobby_is_configured 
         _context = new DefaultHttpContext();
         _context.Request.Path = "/";
 
-        var identity = new ClaimsIdentity(
-            [new Claim("sub", "user-123")], "aad");
+        var identity = new ClaimsIdentity([new Claim("sub", "user-123")], "aad");
         _context.User = new ClaimsPrincipal(identity);
-
         _context.Request.Headers.Cookie = $"{Cookies.InviteToken}=pending-invite-token";
     }
 
     async Task Because() => await _middleware.InvokeAsync(_context);
 
-    [Fact] void should_not_call_next() => _nextCalled.ShouldBeFalse();
-    [Fact] void should_redirect_to_lobby() => _context.Response.Headers.Location.ToString().ShouldEqual(LobbyUrl);
+    [Fact] void should_call_next() => _nextCalled.ShouldBeTrue();
     [Fact] void should_delete_invite_cookie() => _context.Response.Headers.SetCookie.ToString().ShouldContain(Cookies.InviteToken);
+    [Fact] void should_not_set_lobby_redirect_when_exchange_fails() => _context.Items.ContainsKey(InviteMiddleware.LobbyRedirectUrlItemKey).ShouldBeFalse();
 
     static IOptionsMonitor<C.Authentication> CreateEmptyAuthConfig()
     {
         var monitor = Substitute.For<IOptionsMonitor<C.Authentication>>();
         monitor.CurrentValue.Returns(new C.Authentication());
         return monitor;
+    }
+
+    class ThrowingHttpMessageHandler : HttpMessageHandler
+    {
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken) =>
+            throw new HttpRequestException("Simulated exchange failure");
     }
 }
