+++ Source/AuthProxy.Specs/Invites/for_InviteMiddleware/when_authenticated_user_has_pending_tenant_invite/and_tenant_matches.cs
@@ -3,10 +3,14 @@
 
 using System.Net;
 
-namespace Cratis.AuthProxy.Invites.for_InviteMiddleware;
+namespace Cratis.AuthProxy.Invites.for_InviteMiddleware.when_authenticated_user_has_pending_tenant_invite;
 
-public class when_authenticated_user_has_pending_invite_and_exchange_returns_non_success_status : Specification
+public class and_tenant_matches : Specification
 {
+    const string LobbyUrl = "http://lobby-service/";
+    const string TenantClaimType = "tenant_id";
+    const string TenantId = "tenant-matched";
+
     InviteMiddleware _middleware;
     DefaultHttpContext _context;
     bool _nextCalled;
@@ -14,15 +18,22 @@ public class when_authenticated_user_has_pending_invite_and_exchange_returns_non
     void Establish()
     {
         var tokenValidator = Substitute.For<IInviteTokenValidator>();
+        tokenValidator.TryGetClaim(Arg.Any<string>(), TenantClaimType, out Arg.Any<string>())
+            .Returns(x =>
+            {
+                x[2] = TenantId;
+                return true;
+            });
 
         var config = new C.AuthProxy
         {
             Invite = new C.Invite
             {
                 ExchangeUrl = "http://studio/internal/invites/exchange",
+                TenantClaim = TenantClaimType,
                 Lobby = new C.Service
                 {
-                    Frontend = new C.ServiceEndpoint { BaseUrl = "http://lobby-service/" }
+                    Frontend = new C.ServiceEndpoint { BaseUrl = LobbyUrl }
                 }
             }
         };
@@ -31,7 +42,7 @@ public class when_authenticated_user_has_pending_invite_and_exchange_returns_non
 
         var httpClientFactory = Substitute.For<IHttpClientFactory>();
         httpClientFactory.CreateClient(Arg.Any<string>()).Returns(
-            new HttpClient(new FakeHttpMessageHandler(HttpStatusCode.BadRequest)));
+            new HttpClient(new FakeHttpMessageHandler(HttpStatusCode.OK)));
 
         _middleware = new InviteMiddleware(
             _ =>
@@ -49,16 +60,19 @@ public class when_authenticated_user_has_pending_invite_and_exchange_returns_non
         _context = new DefaultHttpContext();
         _context.Request.Path = "/";
 
-        var identity = new ClaimsIdentity([new Claim("sub", "user-123")], "aad");
+        var identity = new ClaimsIdentity(
+            [new Claim("sub", "user-123")], "aad");
         _context.User = new ClaimsPrincipal(identity);
+
         _context.Request.Headers.Cookie = $"{Cookies.InviteToken}=pending-invite-token";
+        _context.Items[TenancyMiddleware.TenantIdItemKey] = TenantId;
     }
 
     async Task Because() => await _middleware.InvokeAsync(_context);
 
     [Fact] void should_call_next() => _nextCalled.ShouldBeTrue();
+    [Fact] void should_not_redirect_to_lobby() => _context.Response.Headers.Location.ToString().ShouldNotContain(LobbyUrl);
     [Fact] void should_delete_invite_cookie() => _context.Response.Headers.SetCookie.ToString().ShouldContain(Cookies.InviteToken);
-    [Fact] void should_not_redirect_to_lobby_when_exchange_fails() => _context.Response.Headers.Location.ToString().ShouldEqual(string.Empty);
 
     static IOptionsMonitor<C.Authentication> CreateEmptyAuthConfig()
     {
