+++ Source/AuthProxy.Specs/Invites/for_InviteMiddleware/when_valid_invite_token_is_presented/and_no_providers_are_configured.cs
@@ -1,9 +1,9 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.AuthProxy.Invites.for_InviteMiddleware;
+namespace Cratis.AuthProxy.Invites.for_InviteMiddleware.when_valid_invite_token_is_presented;
 
-public class when_valid_invite_token_is_presented_and_no_providers_are_configured : Specification
+public class and_no_providers_are_configured : Specification
 {
     InviteMiddleware _middleware;
     DefaultHttpContext _context;
