+++ Source/AuthProxy.Specs/for_TenancyMiddleware/when_tenant_resolution_fails/and_invite_path_is_requested_with_lobby_configured.cs
@@ -1,9 +1,9 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.AuthProxy.for_TenancyMiddleware;
+namespace Cratis.AuthProxy.for_TenancyMiddleware.when_tenant_resolution_fails;
 
-public class when_tenant_resolution_fails_and_auth_path_is_requested_with_lobby_and_resolutions_configured : Specification
+public class and_invite_path_is_requested_with_lobby_configured : Specification
 {
     const string LobbyUrl = "http://lobby-service/";
 
@@ -15,10 +15,6 @@ public class when_tenant_resolution_fails_and_auth_path_is_requested_with_lobby_
     {
         var config = new C.AuthProxy
         {
-            TenantResolutions =
-            [
-                new C.TenantResolution { Strategy = C.TenantSourceIdentifierResolverType.Host }
-            ],
             Invite = new C.Invite
             {
                 Lobby = new C.Service
@@ -42,17 +38,15 @@ public class when_tenant_resolution_fails_and_auth_path_is_requested_with_lobby_
             optionsMonitor,
             tenantResolver,
             Substitute.For<ITenantVerifier>(),
-            Substitute.For<IIdentityDetailsResolver>(),
             Substitute.For<IErrorPageProvider>(),
             Substitute.For<ILogger<TenancyMiddleware>>());
 
         _context = new DefaultHttpContext();
-        _context.Request.Path = WellKnownPaths.Providers;
+        _context.Request.Path = "/invite/some-token";
     }
 
     async Task Because() => await _middleware.InvokeAsync(_context);
 
     [Fact] void should_call_next() => _nextCalled.ShouldBeTrue();
     [Fact] void should_not_redirect_to_lobby() => _context.Response.Headers.Location.ToString().ShouldEqual(string.Empty);
-    [Fact] void should_not_return_401() => _context.Response.StatusCode.ShouldEqual(StatusCodes.Status200OK);
 }
